bug: fixed text overflowing into other containers
The text was overflowing because the format function was replacing
normal spaces with &nsp which was causing the text to overflow all
containers. I also changed from using modals to a hover over to display
movie details so as to provide a better user experience. I also coded
as closely to PEP 8 as possible.